Radical removal of all unneeded code
