--- UK_Advanced_Cryptics_Dictionary_Licensing_Information.txt
@@ -1,12 +1,0 @@
-The following license applies to the wordlist (for model testmain.word) included in testapp/testmain/fixtures/initial_data.json.
-
-Src: http://dreamsteep.com/projects/the-english-open-word-list.html
-======================================================================================
-
-UK Advanced Cryptics Dictionary Licensing Information:
-
-Copyright © J Ross Beresford 1993-1999. All Rights Reserved.
-
-The following restriction is placed on the use of this publication: if the UK Advanced Cryptics Dictionary is used in a software package or redistributed in any form, the copyright notice must be prominently displayed and the text of this document must be included verbatim.
-
-There are no other restrictions: I would like to see the list distributed as widely as possible.
