--- pytest.ini
@@ -1,10 +1,0 @@
-[pytest]
-norecursedirs=env testapp docs
-addopts = --tb=short --pep8 --flakes --isort -rxs
-DJANGO_SETTINGS_MODULE=tests.testapp.settings
-pep8maxlinelength=139
-pep8ignore=
-    runtests.py ALL
-flakes-ignore=
-    django_select2/__init__.py UnusedImport
-    django_select2/fields.py UnusedImport
