+++ requirements_dev.txt
@@ -1,10 +1,9 @@
 pytest
-pytest-pep8
-pytest-flakes
 pytest-django
-pytest-isort
 pep257
 selenium
 model_mommy
 isort
 requests
+flake8
+pep8-naming
