+++ setup.cfg
@@ -1,0 +1,23 @@
+[pytest]
+addopts = --tb=short -rxs
+DJANGO_SETTINGS_MODULE=tests.testapp.settings
+
+[flake8]
+max-line-length = 120
+max-complexity = 10
+statistics = true
+show-source = true
+exclude = docs,runtests.py,setup.py,env
+
+[pep257]
+ignore = D100,D101,D102,D103
+explain = true
+count = true
+
+[isort]
+atomic = true
+multi_line_output = 5
+line_length = 79
+skip = manage.py,docs
+known_first_party = django_select2
+combine_as_imports = true
