--- testapp/testapp/templates/500.html
@@ -1,7 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-</head>
-<body>
-    Error 500
-</body>
