--- testapp/testapp/templates/formget.html
@@ -1,22 +1,0 @@
-{% load staticfiles %}
-{% load django_select2_tags %}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<script src="{{ STATIC_URL }}jquery-1.7.2.min.js"></script>
-	{% import_django_select2_js %}
-	{% import_django_select2_css %}
-	<!-- For testing importing it again, but with another tag and light=1 -->
-	{% import_django_select2_js_css light=1 %}
-</head>
-<body>
-<form method="get" action="">
-	<table>
-		{{form}}
-	</table>
-	<input type="submit" value="Submit Form"/>
-</form>
-{% for result in results %}
-    <p> {{ result.name}} </p>
-{% endfor %}
-</body>
