--- testapp/testapp/templates/index.html
@@ -1,19 +1,0 @@
-{% load url from future %}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-</head>
-<body>
-<h1>Manual Tests</h1>
-<ul>
-	<li><a href="{% url 'test_single_value_model_field' %}">Test single selection model fields</a></li>
-	<li><a href="{% url 'test_multi_values_model_field' %}">Test multi selection model fields</a></li>
-	<li><a href="{% url 'test_mixed_form' %}">Test mixed form. All fields' search must return their own results, not other fields'.</a></li>
-	<li><a href="{% url 'test_init_values' %}">Test that initial values are honored in unbound form</a></li>
-	<li><a href="{% url 'test_list_questions' %}">Test tagging support</a></li>
-	<li><a href="{% url 'test_auto_multivalue_field' %}">Test multi value auto model field.</a></li>
-	<li><a href="{% url 'test_auto_heavy_perf' %}">Test performance. Issue#54.</a></li>
-    <li><a href="{% url 'test_get_search_form' %}">Test a search form using GET. Issue#66.</a></li>
-    <li><a href="{% url 'test_issue_73' %}">Test issue#73.</a></li>
-</ul>
-</body>
