--- testapp/testapp/templates/list.html
@@ -1,26 +1,0 @@
-{% load url from future %}
-<!DOCTYPE html>
-<html lang="en">
-<head>
-</head>
-<body>
-<h2>{{title}}</h2>
-{% if create_new_href != '' %}
-	<a href="{% url create_new_href %}">Create New</a>
-	<br/>
-{% endif %}
-<p>Auto Tags</p>
-<ul>
-    {% for e in object_list %}
-        <li><a href="{% url href e.id %}">{{ e }}</a></li>
-    {% endfor %}
-</ul>
-{% if href_non_auto %}
-<p>Non-Auto Tags</p>
-<ul>
-    {% for e in object_list %}
-        <li><a href="{% url href_non_auto e.id %}">{{ e }}</a></li>
-    {% endfor %}
-</ul>
-{% endif %}
-</body>
