+++ tests/test_fields.py
@@ -1,10 +1,15 @@
 # -*- coding:utf-8 -*-
 from __future__ import print_function, unicode_literals
 
+import json
+
 import pytest
 from django.core.urlresolvers import reverse
-from model_mommy import mommy
+from django.utils.encoding import smart_text
 from selenium.common.exceptions import NoSuchElementException
+
+from django_select2.types import NO_ERR_RESP
+from tests.testapp.forms import AlbumForm, ArtistForm
 
 
 class ViewTestMixin(object):
@@ -15,11 +20,6 @@ class ViewTestMixin(object):
         assert response.status_code == 200
 
 
-@pytest.fixture
-def genres(db):
-    mommy.make('testapp.Genre', _quantity=100)
-
-
 class TestAutoModelSelect2TagField(object):
     url = reverse('single_value_model_field')
 
@@ -28,3 +28,23 @@ class TestAutoModelSelect2TagField(object):
         with pytest.raises(NoSuchElementException):
             error = driver.find_element_by_xpath('//body[@JSError]')
             pytest.fail(error.get_attribute('JSError'))
+
+    def test_form(self):
+        form = ArtistForm()
+        assert form
+
+
+class TestAutoModelSelect2Field(object):
+    def test_form(self, client, artists):
+        artist = artists[0]
+        form = AlbumForm()
+        assert form.as_p()
+        field_id = form.fields['artist'].widget.widget_id
+        url = reverse('django_select2_central_json')
+        response = client.get(url, {'field_id': field_id, 'term': artist.title})
+        assert response.status_code == 200
+        data = json.loads(response.content.decode('utf-8'))
+        assert data['results']
+        assert {'id': artist.pk, 'text': smart_text(artist)} in data['results']
+        assert data['more'] is False
+        assert data['err'] == NO_ERR_RESP
