+++ tests/test_widgets.py
@@ -12,8 +12,8 @@ from model_mommy import mommy
 from selenium.common.exceptions import NoSuchElementException
 from six import text_type
 
-from django_select2 import AutoHeavySelect2Widget
 from django_select2.cache import cache
+from django_select2.forms import AutoHeavySelect2Widget
 from tests.testapp.models import Genre
 
 
@@ -21,12 +21,12 @@ class TestWidgets(object):
     url = ""
 
     def test_is_hidden_multiple(self):
-        from django_select2.widgets import HeavySelect2MultipleWidget
+        from django_select2.forms import HeavySelect2MultipleWidget
         new_widget = HeavySelect2MultipleWidget(data_url="/")
         assert new_widget.is_hidden is False
 
     def test_is_hidden(self):
-        from django_select2.widgets import HeavySelect2Widget
+        from django_select2.forms import HeavySelect2Widget
         new_widget = HeavySelect2Widget(data_url="/")
         assert new_widget.is_hidden is False
